App/requirements.txt
alembic==1.0.8
APScheduler==3.6.1
autopep8==1.4
certifi==2018.10.15
chardet==3.0.4
clarifai==2.4.1
click==6.7
configparser==3.5.0
dominate==2.3.1
EasyProcess==0.2.3
emoji==0.5.1
Flask==1.0.2
Flask-Migrate==2.4.0
flask-nav==0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.17.1
googleapis-common-protos==1.5.5
grpcio==1.16.1
idna==2.7
instapy-chromedriver==2.44
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.8
MarkupSafe==1.0
MeaningCloud-python==1.1.1
Pebble==4.3.9
plyer==1.3.1
protobuf==3.6.1
psutil==5.4.8
pycodestyle==2.4.0
pycryptodome==3.6.6
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
PyVirtualDisplay==0.2.1
PyYAML==3.13
regex==2018.11.22
requests==2.20.1
selenium==3.141.0
six==1.11.0
SQLAlchemy==1.2.12
tzlocal==1.5.1
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
